+++ ale/formatters/formatter.py
@@ -140,7 +140,7 @@ def to_isd(driver):
     instrument_pointing['constant_rotation'] = constant_rotation.rotation_matrix().flatten()
     isd['instrument_pointing'] = instrument_pointing
 
-    # interiror orientation
+    # interior orientation
     isd['naif_keywords'] = driver_data["naif_keywords"]
 
     if isinstance(driver,LineScanner) or isinstance(driver, Framer) or isinstance(driver, PushFrame):